Toward integration of general.sh in backup_sql.sh

+++ backup_sql.sh
@@ -150,14 +150,20 @@ function dumpBase()
     fi
 }
 
+#######
 # Main
-
+########
 if [ ! \( -d $BAK_DIR -a -w $BAK_DIR \) ]; then
   fileLogger "$KO ERR dossier `basename $BAK_DIR` inaccessible"
   exit 1
 fi
+if [ ! \( -d $BAK_DIR_PUB -a -w $BAK_DIR_PUB \) ]; then
+  fileLogger "$KO ERR dossier `basename $BAK_DIR_PUB` inaccessible"
+  exit 1
+fi
+
 if [ ! -f $BAK_DIR_PUB/.htaccess ]; then
-  fileLogger "$KO ERR fichier .htaccess inaccessible"
+  fileLogger "$KO ERR fichier .htaccess inaccessible dans  `basename $BAK_DIR_PUB`"
   rm -f $BAK_DIR_PUB/$SQL_BASE1.sql.zip $BAK_DIR_PUB/$SQL_BASE2.sql.zip
   exit 1
 fi
